update 10 - fractal births and rotations
Transcripts
00:04
We are drawing here is not a fractal yet. It's just a line that has been translated and rotated around to create a star shape. Mathematical fractals are infinitely complex patterns that are self-similar across different scales. It's a repeating pattern of similar shapes that you can zoom into closer and closer to get more and more details. In creative coding, we can't have our fractals to be infinitely complex because then it would take JavaScript infinite time to draw them. We can give them layers of complexity until we scaled to
00:35
a certain small value beyond which we can't really see any difference with naked eye anyway, because individual lines are so small, the fractal shape we will draw today will be made out of lines. And what I'm going for is something inspired by H-tree fractal, where you have three lines that make the base h letter shape that branches into four smaller h letter shapes at each ending and at each ending of these smaller ones, there are even smaller. In mathematical fractal, we can keep zooming closer and closer and see more detail as this pattern repeats
01:07
itself on infinitely smaller and smaller scales. For our effect, we will define some depth. Let's say we want each line to be split into smaller lines five times, and then our fractal shape is complete. Let me show you what I mean. We will draw our fractal, making this private drawLine method call itself over and over. In programming, this is called recursion. Recursion is a process in which a function calls itself directly or indirectly. Each time draw line calls itself, it will draw a line. It will transform, rotate,
01:38
and scale canvas and it will call itself to draw that same line, but this time transformed, scaled, and rotated, and then it will call itself again. We will add a condition to define how many times this process will repeat itself, depending on how much detail, how much depth we want in our final fractal shape. I will show you exactly how that works and what it looks like now. I start by isolating any changes I make to canvas state. So I will wrap my code in save and restore methods. Inside our rotate canvas by one radian.
02:09
And I will make drawLine call itself. I need to pass it context argument. Be careful here, if you run this code, it will create an infinite loop. Most modern browsers could deal with it these days and they will not crash. We need to create a condition that will end recursion at some point that will prevent a drawLine from colon itself and if further, when a certain limit is reached, I do it by passing it a second argument called level. I create an if statement and I check if this level variable that will be passed as an argument is more than this dot max level.
02:41
If it is, use return statement like this. When a return statement is used in a function body, the execution of the function is stopped. I need to create max level property up here. Let's say I want drawLine to call itself two times. Our fractal will have a depth of two levels. When drawLine calls itself, I can see up here that it expects two arguments, contexts and level it every time drawLine calls itself, I increase that level argument by plus one. I run the code and we still get an infinite loop.
03:12
It's because that level variable is undefined. I need to give it an initial value when I call drawLine for the very first time, I set the initial level to 0. So we call drawLine private method and we pass it context to specify which canvas we wanted to draw on and we pass it initial level 0, this code runs to draw a line. And then through a programming principle called recursion, the function calls itself passing along that same context variable and increasing level variable by plus one.
03:43
So at this point, level is one and max level is two, as we defined it up online 19, this check will be false and it will rotate canvas and draw another line, increasing level by one. This loop will be repeating itself, drawing more and more lines until level reaches max level. Every time we draw another line, we can also scale canvas down. I scale it down times 0.9 horizontally and vertically. And now each line is 90% size of the previous one.
04:14
If I want to create a branch shape, I also want to move the point from which the new line grows from its parent line. I will do it by translating canvas state. As we explained earlier, this will push coordinates 0 0 by that value, and it will also move rotation and scale in center point with it as a side effect, for example, look at what happens if between each line I translate by 50 pixels horizontally and 0 vertically, I can push the new line
04:44
along its parent line using translate. What if I want the new line to grow from the end of its parent line, I can do that by translating horizontally by the entire length of that line. So here I used this.size property from line 17. I created kind of a hook shape. I can scale it down a bit. I can change the size of individual segments here on line 17. Now if I increase max level, I allow drawLine to call itself more times increasing in the amount of lines that make up our final fractal shape.
05:14
You can see that with these settings, we quickly reach a point where new lines are scaled down to a point where they become barely visible. With mathematical fractals, this will be infinite and we would be able to zoom closer and closer and see more and more lines. In creative coding, we only care about the part we can see, also we don't have unlimited computing power to create infinite shapes like that. So at any point in this tutorial, if the shapes become too complex and rendering time takes too long on your computer, the easiest way to make them render faster is to come here to
05:46
max level property and set it to a lower value decreasing the depth of our fractal. I will set it to five and I set the scale to 0.7. If I change this angle value in radians, I pass rotate method. I change the angle between new branch and its parent branch. This can create really cool shapes and movement I showed in the intro. So this is our very simple fractal shape. We made a fractal hook. I can now use the code we wrote here on line 26 to turn it into a shape with radial symmetry.
06:17
This for-loop draws a fractal branch, each time it runs and it distributes these branches evenly in a circular area. The number of branches depends on this.size variable from line 18, Let's increase it to 2 3 4 5 6. I hope you can see we are getting closer to something interesting. I go back to one branch for now so that we can clearly see what we are doing. I will also set max level to 1 for now.