Fractal Snowflakes
Transcripts
00:04
The way our code works right now, it draws a line, rotates in a positive direction, and draws another line again. This way we draw these curved hooks. What if I want more symmetry? Let's say I want one branch to grow to the left and one to the right. I want to create kind of a fork that can give them interesting snowflake, tree branch or crystal shapes. Let me show you. We will have two branches growing out from the main parent branch. One will be splitting the left, one to the right.
00:35
Both of these branches will be the same scale, so I move, translate and scale outside this save call here. And I wrap this entire block in another pair of save and restore to make sure our translates, scales and rotates don't overflow outside this loop of drawLine. I hope you don't find it confusing that we have multiple save and restore calls here. Think of the save and restore in pairs. They are kind of nested in each other. So each restore knows which one is its associated save point. So it knows what state of canvas to restore back to.
01:08
This restore will work with this save. This restore is paired with this save. I hope it makes sense. Save pushes the current state on top of the stack, restore pops the last added state of the top of the stack. It will become more intuitive if you used it for awhile. I copy this code block one more time. And the only change will be that we pass rotate method and negative angle. I put some spaces here for clarity. When we call draw line, we draw a line. We translate the position 0 0 along the branch.
01:40
So it's on the other end. We rotate canvas by a positive angle value and call drawLine to draw this branch. And here we rotate it by a negative value to draw this one. Right now, I'm hard-coding scale 0.7. Let's make it into a class property. I say this.scale is 0.7 and I replace it here and here. Now I can change scaling of our fractal levels here on line 20. It will do the same thing for angle value. We pass to rotate method.
02:10
It will affect how spread out the final fractal shape is. So I will call it this.spread. You can also call it angle or something if you want. Initially I set it to one radian, which is around 57 degrees. I replace the hard-coded angle value here on line 46 and on line 51, I will use the same value but negative to make sure these two smaller branches split out in both directions away from the parent branch. If I change the angle value here you can see exactly how it affects the relationship
02:41
between child and parent branch. I add more sides. I can change any value here to change what the final shape looks like. Here on line 35, we limit the number of drawLine method calls itself. In this case, each time it runs, it calls itself twice to create two smaller versions of the same line growing from the tips. Look what happens if I increase max level by one, we will add one more level of recursion and each smaller branch will have two even smaller branches growing from it.
03:12
If I add one more, the smallest branches will each have two even smaller ones. I hope you can see what is happening. I will change the angle value and it will become more obvious that we have our base level shape and three more fractal levels branching from it. We have a branch that splits into two smaller branches. And each time we add max level, this process is repeated on smaller and smaller scale. Basic principle of fractals. I make the entire shape smaller by setting the initial segment size to 20% of canvas width.
03:45
I set line width to 30 pixels. Let's take it one step further. Let's say this is the main branch and it has two smaller branches coming out from it here and here. What if I want to have more branches? For example, I want another pair of coming from somewhere here and here. Let's turn to a number of branches into a class property. Initially I set it to one. I go inside our private drawLine method. And after I draw the main initial line I create a for loop. i is 0, as long as i is less than this dot
04:15
branches increase i by one. Now, I take this entire code and I put it inside that for loop. This code translates coordinates 0 0 along the parent branch to its end and scales canvas down a bit. Then it draws this line here and this line here. I need to move this save outside the for-loop because it's associated restore is also outside the for-loop here. I want one set of branches somewhere here, and the other set here, the code that determines where the next branches grow out of is this translate.
04:48
Instead of translating all the way to the end of parent branch, I will translate to this dot size divided by branches and this whole thing in brackets times i. That way if we have, let's say two sets of branches, the first set will be size divided by two times a 0 here, and the second pair will be size divided by two times one. So here we just have one set now because this dot branches is set to 1. If I want to move it to the other side of the main branch,
05:18
I set this value to this dot size minus like this. I appreciate that this code in line 44 could be a little hard to understand at first. It will become more clear as we play with the code and see how the final fractal shape reacts. I set branches to two and I can see it did something I did not expect. Let's see what that looks like in a more complex shape by adding more sides. I actually like this. The reason the branches are not spread out along the main branch but are pushed kind of outside is because I'm calling this save and
05:49
restore only ones outside of the for-loop. I actually want to save and restore each time we draw a pair of branches so that this translate online 44 can correctly climb along the parent branch. I move save and restore inside the for-loop. And now it's doing what I intended. If I scale it down, you can see we have two sets of branches. One set growing from this point and another set growing from this point, I adjust some of these values to get a better size. Now we are making really nice fractal snowflakes.